Refactor state and fix paths
